custom message coming on wrong authorization with proper error code

+++ Incident-mgmt/src/main/java/com/bb/Incident/mgmt/controller/IncidentController.java
@@ -30,7 +30,7 @@ public class IncidentController {
 
     @Operation(summary = "Get all incidents", description = "Returns a paginated list of all incidents")
     @GetMapping
-    @PreAuthorize("hasAuthority('incident.get')")
+//    @PreAuthorize("hasAuthority('incident.get')")
     public Map<String, Object> getAllIncidents(
             @PageableDefault(size = 5) Pageable pageable,
             @RequestParam(required = false) String incidentType,
@@ -39,9 +39,9 @@ public class IncidentController {
             @RequestParam(required = false) String priority
     ) {
 
-//        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.get"))) {
-//            throw new CustomAccessDeniedException("You need special permissions to get incidents.");
-//        }
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.get"))) {
+            throw new CustomAccessDeniedException("You need special permissions to get incidents.");
+        }
 
         Page<IncidentResponse> page = incidentService.getAllIncidents(pageable, incidentType, severity, state, priority);
         Map<String, Object> response = new HashMap<>();
@@ -62,22 +62,37 @@ public class IncidentController {
 
     @Operation(summary = "Get incident with UUID", description = "Returns the incident with UUID")
     @GetMapping("/{uuid}")
-    @PreAuthorize("hasAuthority('incident.get')")
+//    @PreAuthorize("hasAuthority('incident.get')")
     public IncidentResponse getIncidentById(@PathVariable String uuid) {
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.get"))) {
+            throw new CustomAccessDeniedException("You need special permissions to get the incident with UUID: " + uuid);
+        }
+
         return incidentService.getIncidentByUuid(uuid);
     }
 
     @Operation(summary = "Get Open Incidents", description = "Returns a list of all Open Incidents")
     @GetMapping("/open")
-    @PreAuthorize("hasAuthority('incident.get')")
+//    @PreAuthorize("hasAuthority('incident.get')")
     public List<IncidentResponse> getOpenIncidents() {
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.get"))) {
+            throw new CustomAccessDeniedException("You need special permissions to get the open incidents.");
+        }
+
         return incidentService.getOpenIncidents();
     }
 
     @Operation(summary = "Create Incident", description = "Creates an incident")
     @PostMapping
-    @PreAuthorize("hasAuthority('incident.create')")
+//    @PreAuthorize("hasAuthority('incident.create')")
     public IncidentResponse createIncident(@RequestBody Incident incident) {
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.create"))) {
+            throw new CustomAccessDeniedException("You need special permissions to create an incident.");
+        }
+
         Incident createdIncident = incidentService.createIncident(incident);
         System.out.println("Created Incident, " + createdIncident);
         return incidentService.convertToResponse(createdIncident);
@@ -85,29 +100,49 @@ public class IncidentController {
 
     @Operation(summary = "Update an incident with UUID", description = "Updates an incident with UUID")
     @PutMapping("/{uuid}")
-    @PreAuthorize("hasAuthority('incident.update')")
+//    @PreAuthorize("hasAuthority('incident.update')")
     public IncidentResponse updateIncident(@PathVariable String uuid, @Valid @RequestBody UpdateIncidentRequest updateIncidentRequest) {
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.update"))) {
+            throw new CustomAccessDeniedException("You need special permissions to update the incident with UUID: " + uuid);
+        }
+
         return incidentService.updateIncident(uuid, updateIncidentRequest);
     }
 
     @Operation(summary = "Deletes an incident with UUID", description = "Deletes an incident with UUID")
     @DeleteMapping("/{uuid}")
-    @PreAuthorize("hasAuthority('incident.delete')")
+//    @PreAuthorize("hasAuthority('incident.delete')")
     public void deleteIncident(@PathVariable String uuid) {
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.delete"))) {
+            throw new CustomAccessDeniedException("You need special permissions to delete the incident with UUID: " + uuid);
+        }
+
         incidentService.deleteIncident(uuid);
     }
 
     @Operation(summary = "Resolve an incident", description = "Converts the state of incident to close and updates date resolved")
     @PutMapping("/resolve/{uuid}")
-    @PreAuthorize("hasAuthority('incident.update')")
+//    @PreAuthorize("hasAuthority('incident.update')")
     public IncidentResponse resolveIncident(@PathVariable String uuid) {
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.update"))) {
+            throw new CustomAccessDeniedException("You need special permissions to resolve the incident with UUID: " + uuid);
+        }
+
         return incidentService.resolveIncident(uuid);
     }
 
     @Operation(summary = "Assign a user to incident", description = "Randomly assigns a user to the incident")
     @PutMapping("/assignUser/{uuid}")
-    @PreAuthorize("hasAuthority('incident.update')")
+//    @PreAuthorize("hasAuthority('incident.update')")
     public IncidentResponse assignUser(@PathVariable String uuid) {
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("incident.update"))) {
+            throw new CustomAccessDeniedException("You need special permissions to assign an user to incident with UUID: " + uuid);
+        }
+
         return incidentService.assignUser(uuid);
     }
 }
